add: l_2_16

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -102,8 +102,10 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         $category = Category::query()->find($id);
+        if ($category->posts->count()) {
+            return redirect()->route('admin.categories.index')->with('error', 'Ошибка! У категории есть записи');
+        }
         $category->delete();
-//        Category::destroy($id);
         return redirect()->route('admin.categories.index')->with('success', 'Категория удалена');
     }
 }
